Funcionando show projects una tabla autoincrementable

+++ Talentmixer/src/main/java/com/atenea/talentmixer/models/dao/IprojectDao.java
@@ -2,10 +2,8 @@ package com.atenea.talentmixer.models.dao;
 
 import org.springframework.data.repository.CrudRepository;
 
-import com.atenea.talentmixer.models.entities.Project;
+import com.atenea.talentmixer.models.entity.Project;
 
-
-public interface IprojectDao extends CrudRepository<Project,Long> {
+public interface IprojectDao extends CrudRepository<Project, Integer> {
 
 }
-
